Refactor HeroPage to use HeroNavMenu component
Moved the navigation menu logic from HeroPage to the new HeroNavMenu component. This improves code modularity and reusability by isolating the nav
