+++ frontend-next-migration/src/widgets/HeroNavMenu/index.ts
@@ -1,0 +1,1 @@
+export { default as HeroNavMenu } from './ui/HeroNavMenu';
